now server should be to handle all item stuff
it will generate new item per round with max 3
it will apply the effect (the client didnt know yet rn)

+++ src/Shared/MinimalItems.hpp
@@ -1,8 +1,7 @@
 #pragma once
-#include <cstdint>
 
 struct MinimalItems {
-  uint64_t shared_id; // so we can map to the object version
+  int shared_id; // so we can map to the object version
   int type;      // use ItemType
   bool used;
 };
